[topology] Add common logic to the inject TopologySpec
Right now every operator that references a topology needs to have some
logic to process the retrieved CR Spec. It might be a good idea to hide
this logic (and simplify the code maintenance) to the topology object
itself. By doing this we can limit potential future updates to services
that consume a Topology CR by reference.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>