Optimize EnsureDeletedTopologyRef condition
As a follow up of [1] it is ok to simplify the service operators and
blindly pass instance.Status.lastAppliedTopology to the EnsureServiceTopology
function provided by infra-operator.
The logic that removes the finalizer from an existing TopologyRef (if
lastAppliedTopology != nil) is moved here, resulting in a simplification
of the current operators logic.
This patch resolves yet another tech debt generated by the fact that we
started with a different implementation of lastAppliedTopology that
later has been aligned with the existing TopoRef struct.

[1] https://github.com/openstack-k8s-operators/placement-operator/pull/285#discussion_r1967176124

Signed-off-by: Francesco Pantano <fpantano@redhat.com>